+++ tests/responses/test_result.py
@@ -18,12 +18,12 @@ class TestIntentsResult(unittest.TestCase):
             Result()
 
     def test_result_local(self):
-        result = Result({"a": 1}, True)
+        result = Result({"a": 1}, local=True)
         self.assertEqual(result.data, {"a": 1})
         self.assertEqual(result.local, True)
 
     def test_result_local_false(self):
-        result = Result({"a": 1}, False)
+        result = Result({"a": 1}, local=False)
         self.assertEqual(result.data, {"a": 1})
         self.assertEqual(result.local, False)
 
@@ -33,7 +33,7 @@ class TestIntentsResult(unittest.TestCase):
         self.assertEqual(result.local, True)
 
     def test_dict(self):
-        result = Result({"a": 1}, False, target_device_id="id")
+        result = Result({"a": 1}, local=False, target_device_id="id")
         self.assertEqual(
             {"data": {"a": 1}, "local": False, "targetDeviceId": "id"}, result.dict()
         )
@@ -44,7 +44,7 @@ class TestIntentsResult(unittest.TestCase):
         self.assertTrue(not result)
 
     def test_bool_result_with_secondary_device(self):
-        """[HMPOP-273] TargetDeviceId is nullified if data is empty"""
+        """TargetDeviceId is nullified if data is empty"""
         result = Result({})
         self.assertFalse(result)
         result.target_device_id = "Target"
@@ -60,15 +60,6 @@ class TestIntentsResult(unittest.TestCase):
         result = Result({"a": 0})
         result.update(a=1, b=2)
         self.assertEqual({"a": 1, "b": 2}, result.data)
-
-    # TODO: Do we need this?
-    """ 
-    def test_result_dict_is_serializable(self):
-        #Test `Result.dict` method returns serializable
-        import datetime
-        result = Result({'date': datetime.datetime.now()}).dict()
-        self.assertIsInstance(result['data']['date'], str)
-    """
 
     def test_target_device_id(self):
         """Test target device id"""